notebooks/0_generalcode/requirements.txt
# extra
os
sys
pathlib
tqdm

# used
torch
scikeras
Pillow
pandas
matplotlib

#unused
numpy
seaborn
scikit-learn
scipy
imbalanced-learn
requests
tensorboard
